Administrator Setting Done

+++ Book Shop/src/Functionality.cpp
@@ -201,17 +201,22 @@ int sql::delCat()
 {
 	basic::clear();
 	basic::printTable(table);
-	basic::printColor("\nEnter 'aaa' to exit");
+	basic::printColor("\nEnter 'aaa' to exit", 11);
 	basic::printColor("\nWhich category do you want to remove : ", 2);
 	std::getline(std::cin, user);
 	if (user == "aaa")
 		return 0;
-	command = "DELETE TABLE categories WHERE category = "
+	basic::printColor("\nAre you sure you want to continue (y/n) : ", 8);
+	std::getline(std::cin, name);
+	if (name == "n")
+		return 0;
+	command = "DELETE FROM categories WHERE category = "
 		+ basic::quoting(user)  + ";";
 	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
 	command = "DROP TABLE "+
 		basic::quoting(user) + ";";
 	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
+	return 1;
 }
 
 int sql::changeCatName(const std::string& whichTable)
@@ -233,7 +238,6 @@ int sql::changeCatName(const std::string& whichTable)
 	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
 	return 1;
 }
-
 int sql::editCatData(const std::string& whichTable)
 {
 	home:
@@ -242,19 +246,23 @@ int sql::editCatData(const std::string& whichTable)
 	basic::printTable(table);
 	menu::editDataTable();
 	std::getline(std::cin, user);
-	if (user == "4")
+	if (user == "5")
 		return 0;
+	basic::clear();
+	basic::printTable(table);
 	switch (std::stoi(user))
 	{
 	case 1:
+		addData(whichTable);
 		break;
 	case 2:
+		dataChangeName(whichTable);
 		break;
 	case 3:
+		dataChangeQuantity(whichTable);
 		break;
 	case 4:
-		break;
-	default:
+		dataDel(whichTable);
 		break;
 	}
 	goto home;
@@ -267,19 +275,20 @@ int sql::addData(const std::string& leTable)
 	// Field:
 	// ID, name, quantity, price
 	basic::clear();
+	basic::printColor("You chose table: "+leTable, 8);
 	std::string uName{}, uQuan{}, uPrice{};
 	menu::addDataField();
 	id = farthestID();
 	basic::printColor("\nName: ", 2);
 	std::getline(std::cin, uName);
-	basic::printColor("\nQuantity: ", 2);
+	basic::printColor("Quantity: ", 2);
 	std::getline(std::cin, uQuan);
-	basic::printColor("\nPrice: ", 2);
+	basic::printColor("Price: ", 2);
 	std::getline(std::cin, uPrice);
 	if (uName == "aaa" || uQuan == "aaa" || uPrice == "aaa")
 		return 0;
 	command = "INSERT INTO "
-		+ basic::quoting(leTable) + "VALUES("
+		+ basic::quoting(leTable) + " VALUES("
 		+ basic::quoting(std::to_string(++id)) +","
 		+ basic::quoting(uName) + ","
 		+ basic::quoting(uQuan) + ","
@@ -290,23 +299,45 @@ int sql::addData(const std::string& leTable)
 }
 int sql::dataChangeName(const std::string& whichTable)
 {
-	basic::clear();
-	basic::printTable(table);
-	basic::printColor("\nEnter 'aaa' at id input to exit", 4);
-	basic::printColor("\nPlease select the ID that you want to change: ", 2);
+	menu::reqIdwithExit();
 	std::getline(std::cin, pass);
 	if (pass == "aaa")
 		return 0;
 	basic::printColor("\nPlease enter the new name: ", 2);
 	std::getline(std::cin, user);
 	command = "UPDATE "
-		+ basic::quoting(whichTable) + "SET name = "
-		+ basic::quoting(user) + "WHERE id = "
+		+ basic::quoting(whichTable) + " SET name = "
+		+ basic::quoting(user) + " WHERE id = "
 		+ basic::quoting(pass) + ";";
 	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
 	return 1;
 }
 int sql::dataChangeQuantity(const std::string& leTable)
 {
-
+	menu::reqIdwithExit();
+	std::getline(std::cin, pass);
+	if (pass == "aaa")
+		return 0;
+	basic::printColor("\nPlease enter the new quantity : ", 2);
+	std::getline(std::cin, name);
+	command = "UPDATE "
+		+ basic::quoting(leTable) + " SET quantity = "
+		+ basic::quoting(name) + " WHERE id = "
+		+ basic::quoting(pass) + ";";
+	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
+	return 0;
+}
+int sql::dataDel(const std::string& leTable)
+{
+	menu::reqIdwithExit();
+	std::getline(std::cin, pass);
+	basic::printColor("Are you sure you want to continue (y/n) : ", 8);
+	std::getline(std::cin, name);
+	if (name == "n")
+		return 0;
+	command = "DELETE FROM "
+		+ basic::quoting(leTable) + " WHERE id = "
+		+ basic::quoting(pass) + ";";
+	sqlite3_exec(db, command.c_str(), callback, NULL, NULL);
+	return 1;
 }
